kanal/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.6
django-bulk-update-or-create==0.3.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
google-auth==2.9.1
google-auth-oauthlib==0.5.2
gspread==5.4.0
idna==3.3
install==1.3.5
kombu==5.2.4
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyTelegramBotAPI==4.6.1
pytz==2022.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
telebot==0.0.4
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
